센서 동적으로 받도록 변경, 실험 툴 수정 및 추가, 탭 추가
