+++ algorithm.py
@@ -1,0 +1,8 @@
+import sys
+from PyQt5.QtWidgets import QApplication, QWidget
+
+class Algorithm(QWidget):
+
+    def __init__(self):
+        super().__init__()
+        self.initUI()
